+++ tests/extractorExamples/synthetic/correctSamples/test.php
@@ -9,12 +9,12 @@ $translator
     ->translate('');
 
 // withoutCategory
-$translator->translate('messageId1', array(1, 2));
+$translator->translate('messageId1', [1, 2]);
 $translator->translate('messageId2', [1, 2]);
-$translator->translate('message' . 'Id3', array(1, 2));
+$translator->translate('message' . 'Id3', [1, 2]);
 $translator->translate('message' . 'Id4', [1, 2]);
 
-$translator->translate('message' . 'Id' . 5, array(1, 2));
+$translator->translate('message' . 'Id' . 5, [1, 2]);
 $translator->translate('message' . 'Id' . 6, [1, 2]);
 
 $translator->translate('messageId7');
@@ -32,12 +32,12 @@ $translator->translate('message' . 'Id' . 1 . '6');
 $translator->translate('message' . 'Id' . 1 . '7', ['test' => $translator->translate('messageId' . 18)]);
 
 // With categoryName
-$translator->translate('messageId1', array(1, 2), 'categoryName');
+$translator->translate('messageId1', [1, 2], 'categoryName');
 $translator->translate('messageId2', [1, 2], 'categoryName');
-$translator->translate('message' . 'Id3', array(1, 2), 'categoryName');
+$translator->translate('message' . 'Id3', [1, 2], 'categoryName');
 $translator->translate('message' . 'Id4', [1, 2], 'categoryName');
 
-$translator->translate('message' . 'Id' . 5, array(1, 2), 'categoryName');
+$translator->translate('message' . 'Id' . 5, [1, 2], 'categoryName');
 $translator->translate('message' . 'Id' . 6, [1, 2], 'categoryName');
 
 $translator->translate('messageId7', [], 'categoryName');
@@ -55,12 +55,12 @@ $translator->translate('message' . 'Id' . 1 . '6', [], 'categoryName');
 $translator->translate('message' . 'Id' . 1 . '7', ['test' => $translator->translate('messageId' . 19)], 'categoryName');
 
 // With categoryName and complex params
-$translator->translate('messageId1', array('1', 2), 'categoryName2');
+$translator->translate('messageId1', ['1', 2], 'categoryName2');
 $translator->translate('messageId2', [1, '2'], 'categoryName2');
-$translator->translate('message' . 'Id3', array("1", '2'), 'categoryName2');
+$translator->translate('message' . 'Id3', ["1", '2'], 'categoryName2');
 $translator->translate('message' . 'Id4', ["1", '2'], 'categoryName2');
 
-$translator->translate('message' . "Id" . 5, array(null, 2), 'categoryName2');
+$translator->translate('message' . "Id" . 5, [null, 2], 'categoryName2');
 $translator->translate("message" . 'Id' . 6, [null, 2], 'categoryName2');
 
 $translator->translate('messageId7', ['n' => 1], 'categoryName2');